src/testudo.py
import os
import sys
import json
import yaml
import logging
import argparse
from datetime import datetime

class Testudo:
    def __init__(self):
        self.logger = self._setup_logging()
        self.current_phase = 'initial'
        self.secured_positions = []

    def _setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s'
        )
        return logging.getLogger('Testudo')

    def read_cheatsheet(self, phase):
        try:
            with open(f'cheatsheets/{phase}.md', 'r') as f:
                return self._parse_markdown(f.read())
        except Exception as e:
            self.logger.error(f'Error reading cheatsheet: {e}')
            return []

    def _parse_markdown(self, content):
        commands = []
        for line in content.split('\n'):
            if line.startswith('`') and line.endswith('`'):
                commands.append(line.strip('`'))
        return commands

    def execute_phase(self, phase, target):
        self.logger.info(f'Starting phase: {phase}')
        commands = self.read_cheatsheet(phase)
        results = []

        for cmd in commands:
            try:
                cmd = cmd.format(target=target)
                self.logger.info(f'Executing: {cmd}')
                # Simulation - replace with actual execution
                result = f'Simulated execution of {cmd}'
                results.append({'command': cmd, 'result': result})
            except Exception as e:
                self.logger.error(f'Command failed: {e}')

        return self._generate_report(phase, results)

    def _generate_report(self, phase, results):
        report = {
            'phase': phase,
            'timestamp': datetime.now().isoformat(),
            'results': results
        }

        os.makedirs('reports', exist_ok=True)
        with open(f'reports/{phase}_{datetime.now().strftime("%Y%m%d_%H%M%S")}.json', 'w') as f:
            json.dump(report, f, indent=2)

        return report

    def generate_playbook(self, report):
        playbook = {
            'name': f'Testudo {report["phase"]} playbook',
            'hosts': 'targets',
            'tasks': []
        }

        for result in report['results']:
            playbook['tasks'].append({
                'name': f'Execute: {result["command"]}',
                'command': result['command'],
                'register': 'command_output'
            })

        os.makedirs('playbooks', exist_ok=True)
        filename = f'playbooks/{report["phase"]}_{datetime.now().strftime("%Y%m%d_%H%M%S")}.yml'
        with open(filename, 'w') as f:
            yaml.dump([playbook], f)

        return filename

def main():
    parser = argparse.ArgumentParser(description='Testudo Security Testing Tool')
    parser.add_argument('--target', required=True, help='Target to analyze')
    parser.add_argument('--phase', choices=['recon', 'scan', 'exploit'], 
                       default='recon', help='Phase to execute')
    args = parser.parse_args()

    testudo = Testudo()
    report = testudo.execute_phase(args.phase, args.target)
    playbook = testudo.generate_playbook(report)
    print(f'Generated playbook: {playbook}')

if __name__ == '__main__':
    main()