requirements.txt
pyyaml>=5.1
requests>=2.25.1
ansible>=2.9.0
python-nmap>=0.7.1
colorama>=0.4.4